feat(ui): Improve API Call Widget (#52)
* Remove code duplicates and move API Call Widget to a common ReactJS
component, called `Header`, which is now used in most of our container
components.
* Allow users to dynamically insert their authentication token to allow
them to copy the `curl` command and instantly use it. The token is
already stored in the local storage of the browser and is valid for only
a restricted time period, thus not creating an additional attack surface.