fix(ui): Always show fields that a step applies a transform or filter to
We automatically profile the schema of records by inspecting the results
of the pipeline preview.

If pipeline drops fields or entire records, we previously have not been
able to render fields that defined a transform or filter, making it hard
to users to revert them.

With this fix, we automatically add fields, which define a transform or
filter, to the profiled schema of the records, enabling users to always
change any part of the pipeline, regardless of whether the preview
returns records or not.