fix(streams/inspect): Overwrite `max.poll.records`
The `/streams/:uuid/inspect` endpoint allows to provide the parameter
`limit`, which defines the number of to-be-retrieved records.

Internally, the endpoint is using Kafka's Consumer API for retrieving
the records. Thus, the number of retrieved records is additionally
limited by the consumer property `max.poll.records`.

Unless the stream spec overwrites this property, `max.poll.records`
defaults to `500`.

To enable users to provide any value for the `limit` parameter, we now
overwrite `max.poll.records` when retrieving records and set it to
`value`.