fix(ui/pipeline): Load streams before pipeline (#110)
In some cases, the Settings page of a pipeline in the UI does not show
the streams of the pipeline as selected. This happens when the pipeline
object has been loaded before the stream objects.

This commit fixes this issue by enforcing that streams are retrieved
from the backend before the pipeline.