fix(ui/streams): Do not overwrite field `id`
For the grid view, we use a JavaScript library that requires the
row property `id` filled with an unique identifier.

When inspecting a stream that featured a field with the name `id`, we
have accidentally overwritten this field with the row id from the
JavaScript grid.

This commit fixes this issue such that we no longer overwrite field
values of sample records.