fix(stream-inspect): change message retrieval method (#158)
* fix(stream-inspect): add stream inspect mode flag

refs: #155

* chore: run pre-commit

* fix(stream-inspect): changed consumer seek

* fix(stream-inspect): clean up

* fix(stream-inspect): fixed failing test

switched back to distributed polling for this test. We are testing explicit values and using the distributed approach should always guarantee the existence of the last message.

* fix(stream-inspect): add tests cases

* fix(stream-inspect): remove todo

I opt to set the default to non-distributed, so as to not confuse normal users

* fix(stream-inspect): up jvm heap size

test fix for github actions error

* fix(stream-inspect): remove loggers used for development

* fix(stream-inspect): renamed variable since the "inspect" suffix it is implicit

* chore: run pre-commit

* fix(stream-inspect): renamed ,ethod variables

* fix(stream-inspect): rename variable

* fix(stream-inspect): made the stream inspect tests more meaningful

* fix(stream-inspect): change test case assertion

this test tests that strings could be deserialized. Stopped testing for the occurrence of a certain record

* fix(stream-inspect): stop testing for specific records in SerDe

* fix(stream-inspect): changed longdeserialized test assertion

we aren't explicitly checking for the last sent record anymore