fix(ui/deployments): Get pipelines before deployment
When editing a deployment, perform the API call for loading the
pipelines before performing the API call for loading the deployment, to
make sure that the dropdown box always shows the associated pipeline.

If the call to `/api/v1/pipelines` is performed last, the associated
pipeline is not shown in the dropdown box.