feat(deployments/logs): Support defining tailingLines
Introduce the query parameter `tailingLines` to the
`/deployments/:uuid/logs` endpoint, which is set to `100` by default
and allows users to define the number of tailing log lines that shall be
returned by the API endpoint.

Please note that this commit slightly changes the default behavior of the endpoint, which did
not limit the number of returned log lines in the past, potentially causing issues on both the
client and server side.