feat(api/deployments): Forward health response status
The `/api/v1/deployments/{uuid}/health` endpoint returns the health
status of the underlying Kubernetes Deployment.

So far, the endpoint always returned the HTTP status OK (200) but
included the health information in the response body. This is suboptimal
and does not allow for a straightforward integration of the endpoint
with observability and monitoring products.

This commit changes the behavior of the endpoint as follows:

* If the Kubernetes Deployment is healthy, it returns the HTTP status OK (200).
* If the Kubernetes Deployment is not healthy, it returns the HTTP status INTERNAL SERVER ERROR (500).